FIX: resolved node presisting as a wall after reset
resolved bug that caused a node to stay marked as a 'wall' perminatly.
This would cause a visual bug which would mark a node as a 'wall' but the
algorithm would not 'see' the node as a wall. Additionally, some walls would
stay wall perminatly even thorugh full board reset.